feat #1 진입 라우트 수정
