+++ src/layouts/NavBar.vue
@@ -7,7 +7,7 @@
       <div class="user-box">OOO 님의 가계부</div>
 
       <nav class="menu">
-        <router-link to="/">캘린더</router-link>
+        <router-link to="/home">캘린더</router-link>
         <router-link to="/accounts">내역</router-link>
         <router-link to="/dashboard">분석</router-link>
       </nav>
